feat(ledc): Change channel to uint8_t + add log message

+++ cores/esp32/esp32-hal-ledc.h
@@ -45,7 +45,7 @@ typedef struct {
 
 //channel 0-15 resolution 1-16bits freq limits depend on resolution
 bool        ledcAttach(uint8_t pin, uint32_t freq, uint8_t resolution);
-bool        ledcAttachChannel(uint8_t pin, uint32_t freq, uint8_t resolution, int channel);
+bool        ledcAttachChannel(uint8_t pin, uint32_t freq, uint8_t resolution, uint8_t channel);
 bool        ledcWrite(uint8_t pin, uint32_t duty);
 uint32_t    ledcWriteTone(uint8_t pin, uint32_t freq);
 uint32_t    ledcWriteNote(uint8_t pin, note_t note, uint8_t octave);
